requirements.txt
Flask
redis
kafka-python
Flask-Cors
